Refactor useJobTracker hook
Refactor the useJobTracker hook into smaller, more focused hooks for better code organization and maintainability.
